Extract link handlers to hook
