+++ app/javascript/styles/mastodon-light/variables.scss
@@ -60,7 +60,7 @@ body {
   --dropdown-border-color: #d9e1e8;
   --dropdown-background-color: #fff;
   --modal-border-color: #d9e1e8;
-  --modal-background-color: #fff;
+  --modal-background-color: var(--background-color-tint);
   --background-border-color: #d9e1e8;
   --background-color: #fff;
   --background-color-tint: rgba(255, 255, 255, 80%);
